# apps/backend/docs/openapi/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
host: localhost:8080
info:
  contact:
    email: security@arc-hawk.io
    name: ARC-Hawk Security Team
  description: Enterprise PII Discovery, Classification, and Lineage Tracking Platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://arc-hawk.io/terms
  title: ARC-HAWK-DD API
  version: 3.0.0
paths:
...
  /api/v1/compliance/dpdpa/gaps:
    get:
      consumes:
      - application/json
      description: Returns the raw DPDPA obligation gap data as JSON
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get DPDPA obligation gaps
      tags:
      - compliance
  /api/v1/compliance/dpdpa/report:
    get:
      consumes:
      - application/json
      description: Returns an HTML compliance gap report suitable for print-to-PDF
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Generate DPDPA HTML compliance report
      tags:
      - compliance
  /api/v1/compliance/dpr:
    get:
      consumes:
      - application/json
      description: Returns paginated list of data principal rights requests with optional
        filters
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Filter by request status (PENDING, IN_PROGRESS, COMPLETED, REJECTED)
        in: query
        name: status
        type: string
      - description: Filter by request type
        in: query
        name: type
        type: string
      - description: Maximum records to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Number of records to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List data principal rights requests
      tags:
      - compliance
    post:
      consumes:
      - application/json
      description: Submits a new DPDPA data principal rights request (access, correction,
        erasure, etc.)
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data principal request payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Submit a data principal rights request
      tags:
      - compliance
  /api/v1/compliance/dpr/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of a data principal rights request
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request UUID
        in: path
        name: id
        required: true
        type: string
      - description: Status update payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update data principal request status
      tags:
      - compliance
  /api/v1/compliance/dpr/stats:
    get:
      consumes:
      - application/json
      description: Returns counts by status and overdue compliance flag for DPDPA
        Sec 7
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get data principal request statistics
      tags:
      - compliance
...
securityDefinitions:
  APIKeyAuth:
    description: Tenant API key for service-to-service authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"